refractor : Changes in the PostList Component
Earlier the component had a if-else block which had redundant code,
so made changes in the via the ternary operator to decide the array to
display. This array was decided with the help of ternary operator and
new filtered array was created using the filter and sort methods hence
returning a new array of objects each time.